Extract `xcodeGeneratedFiles` calculation
This will allow us to later perform `createFilesAndGroups()` in parallel with other functions.
