+++ test/fixtures/generator/bwx_targets_spec.json
@@ -312,6 +312,7 @@
                 "tools/generator/src/FilePathResolver.swift",
                 "tools/generator/src/Generator/AddBazelDependenciesTarget.swift",
                 "tools/generator/src/Generator/AddTargets.swift",
+                "tools/generator/src/Generator/CalculateXcodeGeneratedFiles.swift",
                 "tools/generator/src/Generator/ConsolidateTargets.swift",
                 "tools/generator/src/Generator/CreateAutogeneratedXCSchemes.swift",
                 "tools/generator/src/Generator/CreateCustomXCSchemes.swift",
