+++ tools/generator/test/SetTargetDependenciesTests.swift
@@ -20,7 +20,7 @@ final class SetTargetDependenciesTests: XCTestCase {
 
         let consolidatedTargets = Fixtures.consolidatedTargets
 
-        let (pbxTargets, disambiguatedTargets, _) = Fixtures.pbxTargets(
+        let (pbxTargets, disambiguatedTargets) = Fixtures.pbxTargets(
             in: pbxProj,
             directories: directories,
             consolidatedTargets: consolidatedTargets
